requirements.txt
huggingface-hub==0.31.4
imageio==2.37.0
matplotlib==3.10.3
numpy==2.2.6
opencv-python==4.11.0.86
pandas==2.2.3
pillow==11.2.1
PyYAML==6.0.2
roboflow==1.1.64
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
ultralytics==8.3.169
